# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Bill Manager API
  version: 1.0.0-m4
  description: Clean Slate M4 本地 API。所有租户资源均从服务端会话解析 TenantContext。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /material-exports/preview:
    post:
      operationId: previewMaterialExport
      tags: [reimbursements]
      description: 完整读取显式当前行程或固定报销快照，要求 facts.read 和 review.source.read；报销另需 reimbursements.read。无文件内容或内部存储键。
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExportScope" }
      responses:
        "200":
          description: 有界完整清单，不能按管理页截断
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExportManifest" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /material-exports:
    post:
      operationId: prepareMaterialExport
      tags: [reimbursements]
      description: 重新比较预览 hash，最长两分钟逐份校验并准备磁盘 ZIP；每进程最多两个在途包，成功后同会话五分钟内领取一次。
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExportPrepareRequest" }
      responses:
        "201":
          description: 完整 ZIP 已准备，尚未下载
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExportPrepared" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "503":
          description: 并发名额已满、磁盘不可写或服务已关闭；不得认为已生成
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /material-exports/{export_id}/content:
    get:
      operationId: downloadMaterialExport
      tags: [reimbursements]
      description: 同一有效 tenant/user/session 一次性领取；不支持 HEAD、Range 或续传。客户端使用原生附件下载，不将 ZIP 读为 Blob。
      parameters:
        - { name: export_id, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9_-]+$' } }
      responses:
        "200":
          description: 已完整校验的 ZIP；精确 Content-Length 和附件文件名
          content:
            application/zip:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "405": { description: HEAD 不消费临时包 }
  /material-exports/{export_id}:
    delete:
      operationId: cancelMaterialExport
      tags: [reimbursements]
      description: 同一会话取消尚未领取的临时包；已领取的下载由浏览器取消，过期或已领取返回 404。
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - { name: export_id, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9_-]+$' } }
      responses:
        "204": { description: 临时文件句柄已释放 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /session/workspaces:
    post:
      operationId: listLoginWorkspaces
      tags: [session]
      security: []
      description: 只有密码验证成功后才返回 active 工作区，不产生 Session。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceLoginRequest" }
      responses:
        "200":
          description: 显式工作区选择
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceChoices" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /members:
    get:
      operationId: listMembers
      tags: [accounts]
      description: 仅 Owner 读取本租户成员，包含 suspended；按入库时间和用户 ID 稳定分页。
      parameters:
        - $ref: "#/components/parameters/FactCursor"
        - $ref: "#/components/parameters/FactLimit"
      responses:
        "200":
          description: 本页成员
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /members/{user_id}:
    get:
      operationId: getMember
      tags: [accounts]
      description: 仅当前租户管理员按 ID 核对成员，冲突恢复不依赖其仍在当前分页。
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 当前成员版本
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: changeMember
      tags: [accounts]
      description: 仅改变本租户角色/状态并撤销其旧会话；不修改全局账号。
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemberChange" }
      responses:
        "200":
          description: 变更后的成员，版本递增
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /member-invitations:
    get:
      operationId: listMemberInvitations
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/FactCursor"
        - $ref: "#/components/parameters/FactLimit"
      responses:
        "200":
          description: Owner 查看邀请历史，不返回代码或 hash
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createMemberInvitation
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationRequest" }
      responses:
        "200":
          description: 48 小时邀请，代码只在首次成功响应返回；回放 code 为空字符串
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /member-invitations/{invitation_id}:
    get:
      operationId: getMemberInvitation
      tags: [accounts]
      description: 仅当前租户管理员按 ID 核对邀请，不返回代码或其散列。
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 当前邀请版本
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /member-invitations/{invitation_id}/revoke:
    post:
      operationId: revokeMemberInvitation
      tags: [accounts]
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationRevocation" }
      responses:
        "200":
          description: 已撤销的邀请
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /invitations/check:
    post:
      operationId: checkInvitation
      tags: [accounts]
      security: []
      description: 代码仅在请求体传输；无效、过期、撤销或已用统一拒绝。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationCode" }
      responses:
        "200":
          description: 持有代码者核对工作区与身份模式
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationView" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [accounts]
      security: []
      description: 新账号设置密码和姓名；已有账号必须验证旧密码，display_name 必须为空。成功不签发 Session。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationAcceptance" }
      responses:
        "204": { description: 已加入，请显式登录 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /account/password:
    post:
      operationId: changeOwnPassword
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordChange" }
      responses:
        "204": { description: 密码已修改，全部工作区会话已撤销，包括当前会话 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { description: unauthenticated 会话失效；invalid_credentials 当前密码不正确，不应清空表单 }
        "403": { $ref: "#/components/responses/Forbidden" }
  /facts/{fact_type}/{fact_id}/correction:
    parameters:
      - $ref: "#/components/parameters/CorrectionFactType"
      - $ref: "#/components/parameters/CorrectionFactId"
    get:
      operationId: getFactCorrection
      tags: [facts]
      description: Owner/Finance 读取未删除正式单据的当前字段与来源，不重开 Job。
      responses:
        "200":
          description: 当前字段、分配和归属
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CorrectionWorkspace" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: confirmFactCorrection
      tags: [facts]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CorrectionConfirmRequest" }
      responses:
        "200":
          description: 同 ID 原子纠错或同请求回放
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CorrectionResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /facts/{fact_type}/{fact_id}/correction/preview:
    parameters:
      - $ref: "#/components/parameters/CorrectionFactType"
      - $ref: "#/components/parameters/CorrectionFactId"
    post:
      operationId: previewFactCorrection
      tags: [facts]
      description: 校验完整字段、全部活动分配、重复和自动归属；不保存草稿。
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CorrectionRequest" }
      responses:
        "200":
          description: 内容绑定预览；can_confirm 为 false 时不得提交
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CorrectionPreview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /facts/{fact_type}/{fact_id}/correction/history:
    get:
      operationId: getFactCorrectionHistory
      tags: [facts]
      parameters:
        - $ref: "#/components/parameters/CorrectionFactType"
        - $ref: "#/components/parameters/CorrectionFactId"
        - name: before_revision
          in: query
          schema: { type: integer, minimum: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: 修订倒序 keyset 历史；首次确认也在其中
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CorrectionHistoryPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /documents/{document_id}/content:
    get:
      operationId: downloadDocument
      tags: [documents]
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: 鉴权后的原始文件
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
            application/pdf: {}
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sbm_session
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
...
  responses:
    BadRequest:
      description: 输入不合法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 权限或 CSRF 校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在或对当前租户不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 幂等、版本、重复或状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PayloadTooLarge:
      description: 请求载荷超过对应接口限制
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    ExportScope:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind: { type: string, enum: [trip, reimbursement] }
        id: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9_-]+$' }
...